# refactor(cbor): Rewrite of CBOR reader. Faster and simpler.