Rewrite CBOR reader so that it avoids buffering tokens. Instead, output parsed elements as they are read off the input reader. Decode CBOR bytes on our own, instead of relying on another library (which is what required buffering).

Performance is about 3x better. Before:

```
> go test ./dsio/. -bench=.

BenchmarkCBORWriterArrays-2    	    3000	    441960 ns/op
BenchmarkCBORWriterObjects-2   	    3000	    617959 ns/op
BenchmarkCBORReader-2          	     200	   6453132 ns/op
BenchmarkCSVWriterArrays-2     	    1000	   1547526 ns/op
BenchmarkCSVWriterObjects-2    	    1000	   1970756 ns/op
BenchmarkCSVReader-2           	     500	   2300161 ns/op
BenchmarkJSONWriterArrays-2    	    1000	   1890182 ns/op
BenchmarkJSONWriterObjects-2   	     500	   2021165 ns/op
BenchmarkJSONReader-2          	     500	   3043123 ns/op
```

(compare to https://github.com/qri-io/dataset/commit/819375c81ba49a904b97dd1f670e90920b4edbca)

After:

```
BenchmarkCBORWriterArrays-2    	    3000	    431290 ns/op
BenchmarkCBORWriterObjects-2   	    2000	    698920 ns/op
BenchmarkCBORReader-2          	    1000	   1764549 ns/op
BenchmarkCSVWriterArrays-2     	    1000	   1548509 ns/op
BenchmarkCSVWriterObjects-2    	    1000	   1458219 ns/op
BenchmarkCSVReader-2           	    1000	   2008097 ns/op
BenchmarkJSONWriterArrays-2    	    1000	   1556416 ns/op
BenchmarkJSONWriterObjects-2   	    1000	   1562488 ns/op
BenchmarkJSONReader-2          	     500	   2984057 ns/op
```