@b5 2018-12-04
[benchcmp](https://godoc.org/golang.org/x/tools/cmd/benchcmp) can make these comparisons even more delightful so we can see just how much faster things got 😄 
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkCBORWriterArrays-4      156580        154605        -1.26%
BenchmarkCBORWriterObjects-4     225393        227865        +1.10%
BenchmarkCBORReader-4            1599268       585046        -63.42%
BenchmarkCSVWriterArrays-4       530450        510958        -3.67%
BenchmarkCSVWriterObjects-4      517925        508732        -1.77%
BenchmarkCSVReader-4             598331        586055        -2.05%
BenchmarkJSONWriterArrays-4      604918        593041        -1.96%
BenchmarkJSONWriterObjects-4     647054        645048        -0.31%
BenchmarkJSONReader-4            1012359       1008421       -0.39%
```

got this by running:
```
$ go get golang.org/x/tools/cmd/benchcmp
$ git checkout master
$ go test ./dsio/. -bench=. > before.txt
$ git checkout cbor-improve
$ go test ./dsio/. -bench=. > after.txt
$ benchcmp before.txt after.txt
```